refactor : 불필요한 유틸클레스를 제거함
