--- src/main/kotlin/calculator/component/Helper.kt
@@ -1,9 +1,0 @@
-package calculator.component
-
-object Helper {
-    fun <T> arrayDequeOf(vararg t: T): ArrayDeque<T> {
-        return ArrayDeque(
-            t.toList()
-        )
-    }
-}
