+++ src/test/kotlin/calculator/app/StringCalculatorTest.kt
@@ -1,8 +1,6 @@
 package calculator.app
 
-import calculator.component.Formula
-import calculator.component.Helper
-import calculator.fixture.TokenFixture
+import calculator.model.Formula
 import calculator.model.Token
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
@@ -10,13 +8,13 @@ import io.kotest.matchers.shouldBe
 class StringCalculatorTest : StringSpec({
 
     "숫자 하나를 문자열로 입력할 경우 해당 숫자를 반환 해야 한다" {
-        val formula = Formula(Helper.arrayDequeOf(TokenFixture.TOKEN[3]))
+        val formula = Formula(listOf(3))
         val actual = StringCalculator.calculate(formula)
         actual shouldBe 3
     }
 
     "숫자 두개를 입력할 경우 두 숫자의 합을 반환 해야 한다 " {
-        val formula = Formula(Helper.arrayDequeOf(TokenFixture.TOKEN[1], TokenFixture.TOKEN[2]))
+        val formula = Formula(listOf(1, 2))
         val actual = StringCalculator.calculate(formula)
         actual shouldBe 3
     }
